moving internal routes into the correct directory
